# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: UpPay Connect API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List all payments
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: provider
          in: query
          schema:
            type: string
        - name: payment_method
          in: query
          schema:
            type: string
        - name: source_system
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentList"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create a new payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /v1/payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get payment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "404":
          description: Not found
  /v1/payments/{id}/cancel:
    post:
      operationId: cancelPayment
      tags: [payments]
      summary: Cancel a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /v1/payments/{id}/refund:
    post:
      operationId: refundPayment
      tags: [payments]
      summary: Refund a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefundInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /v1/payments/{id}/events:
    get:
      operationId: getPaymentEvents
      tags: [payments]
      summary: Get events for a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentEvent"

  # ─── Customers ────────────────────────────────────────────────────────────────
...
  /v1/customers/{id}/payments:
    get:
      operationId: getCustomerPayments
      tags: [customers]
      summary: Get payments for a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"

  # ─── Integrated Systems ──────────────────────────────────────────────────────
  /v1/systems:
    get:
      operationId: listSystems
      tags: [systems]
      summary: List integrated systems
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IntegratedSystem"
    post:
      operationId: createSystem
      tags: [systems]
      summary: Register a new integrated system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntegratedSystemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegratedSystem"
  /v1/systems/{id}:
    get:
      operationId: getSystem
      tags: [systems]
      summary: Get integrated system by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegratedSystem"
    patch:
      operationId: updateSystem
      tags: [systems]
      summary: Update an integrated system
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntegratedSystemUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegratedSystem"

  # ─── API Keys ────────────────────────────────────────────────────────────────
...
  /v1/subscriptions:
    get:
      operationId: listSubscriptions
      tags: [subscriptions]
      summary: List subscriptions
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: source_system
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionList"
    post:
      operationId: createSubscription
      tags: [subscriptions]
      summary: Create a subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /v1/subscriptions/{id}:
    get:
      operationId: getSubscription
      tags: [subscriptions]
      summary: Get subscription by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /v1/subscriptions/{id}/cancel:
    post:
      operationId: cancelSubscription
      tags: [subscriptions]
      summary: Cancel a subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"

  # ─── Webhooks ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Payment:
      type: object
      required:
        - id
        - amount
        - currency
        - payment_method
        - status
        - provider
        - source_system
        - created_at
      properties:
        id:
          type: string
        source_system_id:
          type: ["string", "null"]
        customer_id:
          type: ["string", "null"]
        provider:
          type: string
        provider_payment_id:
          type: ["string", "null"]
        status:
          type: string
        external_reference:
          type: ["string", "null"]
        amount:
          type: integer
        gross_amount:
          type: ["integer", "null"]
        provider_fee:
          type: ["integer", "null"]
        net_amount:
          type: ["integer", "null"]
        currency:
          type: string
        payment_method:
          type: string
        description:
          type: ["string", "null"]
        installments:
          type: ["integer", "null"]
        qr_code:
          type: ["string", "null"]
        qr_code_base64:
          type: ["string", "null"]
        copy_paste:
          type: ["string", "null"]
        boleto_url:
          type: ["string", "null"]
        boleto_barcode:
          type: ["string", "null"]
        boleto_digitable_line:
          type: ["string", "null"]
        payment_link_url:
          type: ["string", "null"]
        source_system:
          type: string
        callback_url:
          type: ["string", "null"]
        expires_at:
          type: ["string", "null"]
        due_date:
          type: ["string", "null"]
        paid_at:
          type: ["string", "null"]
        cancelled_at:
          type: ["string", "null"]
        refunded_at:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: ["string", "null"]
    PaymentList:
      type: object
      required: [data, total, limit, offset]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    PaymentInput:
      type: object
      required: [amount, currency, payment_method, source_system]
      properties:
        amount:
          type: integer
        currency:
          type: string
        payment_method:
          type: string
        description:
          type: string
        external_reference:
          type: string
        source_system:
          type: string
        callback_url:
          type: string
        expires_in:
          type: integer
        customer_id:
          type: string
    RefundInput:
      type: object
      properties:
        amount:
          type: ["integer", "null"]
    PaymentEvent:
      type: object
      required: [id, payment_id, event, created_at]
      properties:
        id:
          type: string
        payment_id:
          type: string
        event:
          type: string
        old_status:
          type: ["string", "null"]
        new_status:
          type: ["string", "null"]
        provider:
          type: ["string", "null"]
        created_at:
          type: string

    # ─── Customer ────────────────────────────────────────────────────────────────
...
    IntegratedSystem:
      type: object
      required: [id, name, slug, status, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        default_webhook_url:
          type: ["string", "null"]
        allowed_ips:
          type: ["string", "null"]
        environment:
          type: string
        created_at:
          type: string
        updated_at:
          type: ["string", "null"]
    IntegratedSystemInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        default_webhook_url:
          type: string
        allowed_ips:
          type: string
        environment:
          type: string
    IntegratedSystemUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
        default_webhook_url:
          type: string
        allowed_ips:
          type: string
        environment:
          type: string

    # ─── API Key ─────────────────────────────────────────────────────────────────
...
    Subscription:
      type: object
      required: [id, customer_id, amount, payment_method, billing_cycle, status, source_system, created_at]
      properties:
        id:
          type: string
        customer_id:
          type: string
        amount:
          type: integer
        payment_method:
          type: string
        billing_cycle:
          type: string
        description:
          type: ["string", "null"]
        source_system:
          type: string
        external_reference:
          type: ["string", "null"]
        status:
          type: string
        provider:
          type: ["string", "null"]
        provider_subscription_id:
          type: ["string", "null"]
        next_due_date:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: ["string", "null"]
    SubscriptionList:
      type: object
      required: [data, total, limit, offset]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Subscription"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    SubscriptionInput:
      type: object
      required: [customer_id, amount, payment_method, billing_cycle, source_system]
      properties:
        customer_id:
          type: string
        amount:
          type: integer
        payment_method:
          type: string
        billing_cycle:
          type: string
        description:
          type: string
        source_system:
          type: string
        external_reference:
          type: string
        next_due_date:
          type: string

    # ─── Webhook Delivery ────────────────────────────────────────────────────────
...