# lib/api-zod/src/generated/types/paymentList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * UpPay Connect API
 * OpenAPI spec version: 0.1.0
 */
import type { Payment } from './payment';

export interface PaymentList {
  data: Payment[];
  total: number;
  limit: number;
  offset: number;
}