# lib/api-zod/src/generated/types/subscriptionList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * UpPay Connect API
 * OpenAPI spec version: 0.1.0
 */
import type { Subscription } from './subscription';

export interface SubscriptionList {
  data: Subscription[];
  total: number;
  limit: number;
  offset: number;
}